Refactor to get rid of process manager
When import is done with process manager async processes - there is a conflict with BlueSpiceWikiFarm.
So it was decided to switch to sync processing for 4.2.x.
In 4.3 new version of process manager will be used, where issue with BlueSpiceWikiFarm probably will be solved, and import could be switched to async processes again.

[4.2.x only]

ERM29351

Change-Id: I159d61c5a83dfc902443239d516d36c4803cef97