Fix convertion of image dimensions
In Word for images width and height inches are used, but for wikitext we need pixels.
There was a typo in calculations, fixed that.

Also fixed mistake with wrong syntax for cases when there is only width or only height.
"x" symbol should be prepended only for "solo height".

[5.x]

ERM29987

Change-Id: Id2a6ec15b50a40b3b7fac6fc6b6635d9e80911eb